Move utils-thread to api
This allows us to use it in intrumentations.
