+++ dd-trace-api/src/main/java/datadog/trace/internal/executor/DaemonThreadFactory.java
@@ -1,4 +1,4 @@
-package datadog.common.exec;
+package datadog.trace.internal.executor;
 
 import java.util.concurrent.ThreadFactory;
