+++ dd-trace-api/src/test/groovy/datadog/trace/internal/executor/PeriodicSchedulingTest.groovy
@@ -1,4 +1,4 @@
-package datadog.common.exec
+package datadog.trace.internal.executor
 
 import datadog.trace.util.gc.GCUtils
 import datadog.trace.util.test.DDSpecification
