+++ dd-trace-ot/dd-trace-ot.gradle
@@ -34,7 +34,6 @@ dependencies {
   implementation deps.autoservice
 
   compile project(':dd-trace-api')
-  compile project(':utils:thread-utils')
 
   compile deps.opentracing
   compile group: 'io.opentracing.contrib', name: 'opentracing-tracerresolver', version: '0.1.0'
@@ -91,15 +90,6 @@ jar {
 
 shadowJar {
   archiveClassifier = ''
-
-  dependencies {
-    include(project(':utils:thread-utils'))
-  }
-}
-
-// We don't want bundled dependencies to show up in the pom.
-modifyPom {
-  dependencies.removeAll { it.artifactId == "thread-utils" }
 }
 
 jmh {
