--- utils/thread-utils/thread-utils.gradle
@@ -1,12 +1,0 @@
-apply from: "${rootDir}/gradle/java.gradle"
-
-// TODO: add more tests
-excludedClassesCoverage += [
-  'datadog.common.exec*'
-]
-
-dependencies {
-  compile deps.slf4j
-
-  testCompile project(':utils:test-utils')
-}
